Train and interpret a model on the Double Majority dataset (#18)
* update workflows

* train on doublemaj

* update makefile

* update metric calculation to use tensors

* finish experiment

* update readme

* add impossibility argument

* add additional documentation

* change readme

+++ src/miexp/bfuncs.py
@@ -46,7 +46,7 @@ class DoubleMajDataset(Dataset):
         self.num_samples = num_samples
         self.high = high
         self.low = low
-        self.data = torch.randint(0, 2, (self.num_samples, self.N)).type(torch.float)
+        self.data = torch.randint(0, 2, (self.num_samples, self.N)).type(torch.int)
         self.labels = (
             (torch.sum(self.data, dim=1) < (self.N * self.high))
             & (torch.sum(self.data, dim=1) >= (self.N * self.low))
